Whatsapp-Redirect-Bot/requirements.txt
anyio==3.6.2; python_version >= "3.7" and python_full_version >= "3.6.2"
apscheduler==3.6.3; python_version >= "3.7"
asgiref==3.6.0; python_version >= "3.7"
cachetools==4.2.2; python_version >= "3.7" and python_version < "4.0"
certifi==2022.12.7; python_version >= "3.7"
click==8.1.3; python_version >= "3.7"
colorama==0.4.6; python_version >= "3.7" and python_full_version < "3.0.0" and sys_platform == "win32" and platform_system == "Windows" or sys_platform == "win32" and python_version >= "3.7" and python_full_version >= "3.7.0" and platform_system == "Windows"
fastapi==0.75.2; python_full_version >= "3.6.1"
gunicorn==20.1.0; python_version >= "3.5"
h11==0.14.0; python_version >= "3.7"
httptools==0.5.0; python_version >= "3.7" and python_full_version >= "3.5.0"
idna==3.4; python_version >= "3.7" and python_full_version >= "3.6.2"
pydantic==1.10.7; python_version >= "3.7" and python_full_version >= "3.6.1"
python-dotenv==0.20.0; python_version >= "3.5"
python-telegram-bot==13.15; python_version >= "3.7"
pytz-deprecation-shim==0.1.0.post0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7"
pytz==2023.3; python_version >= "3.7"
pyyaml==6.0; python_version >= "3.7"
six==1.16.0; python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.7"
sniffio==1.3.0; python_version >= "3.7" and python_full_version >= "3.6.2"
starlette==0.27.0; python_version >= "3.6" and python_full_version >= "3.6.1"
tornado==6.1; python_version >= "3.7"
typing-extensions==4.5.0; python_version >= "3.7" and python_full_version >= "3.6.1"
tzdata==2023.3; platform_system == "Windows" and python_version >= "3.7" and (python_version >= "3.7" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.7")
tzlocal==4.3; python_version >= "3.7"
uvicorn==0.17.6; python_version >= "3.7"
uvloop==0.17.0; sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy" and python_version >= "3.7"
watchgod==0.8.2; python_version >= "3.7"
websockets==11.0.2; python_version >= "3.7"
